backend/pyproject.toml
[project]
name = "summer-dreams-ota"
version = "0.1.0"
description = "Backend for Summer Dreams OTA"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=47.0.0",
    "dj-rest-auth>=7.2.0",
    "django>=6.0.4",
    "django-allauth>=65.16.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "psycopg>=3.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentry-sdk>=2.58.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".gemini",
    "**/migrations/*",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.14 (but ruff supports up to 3.13)
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "T20", # flake8-print
    "DJ",  # flake8-django
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_[a-zA-Z0-9_]*[a-zA-Z0-9]+?)|args|kwargs)$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "django", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django" = ["django"]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.3.0",
    "django-extensions>=4.1",
    "pre-commit>=4.6.0",
    "ruff>=0.15.12",
]
